Added resources and dependencies submodules.
By splitting the resources directory and the libs directory into their
own seperate repos, we can manage the inevitable bloat of each
repository by isolating the bloat by type.

This has tradeoffs, but the point is to be able to download source code
quickly without the need of pulling in dependencies or resources.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>