Updated NodeActivity and optimized the WebView and its video support.
JavaScript is used for updating the WebView on orientation changes,
font size changes, etc. Videos have thumbnail images displayed. If
the native youtube app is not available, the link will be launched
as if it were a normal link to the youtube video.

Also, the JavaScript and CSS has been isolated into their own files.
Padding is now applied using browser CSS instead of directly onto the
WebView.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>